1、修复品牌分析.xml中内部类解析的问题 2、新增价格监测接口
