+++ Gra-backend/src/main/java/com/yupi/springbootinit/mapper/PriceMonitoringMapper.java
@@ -1,0 +1,9 @@
+package com.yupi.springbootinit.mapper;
+
+import com.yupi.springbootinit.model.vo.PriceMonitoringResult;
+
+import java.util.List;
+
+public interface PriceMonitoringMapper {
+    List<PriceMonitoringResult> getPriceMonitoringData(String regionName);
+}
