+++ Gra-backend/src/main/java/com/yupi/springbootinit/model/vo/BrandAnalysisResult.java
@@ -31,6 +31,7 @@ public class BrandAnalysisResult {
     public static class SpecSales {
         private String specification;
         private double totalOrderAmount;
+
     }
 
     // 不同地区的销量数据内部类
