+++ Gra-backend/src/main/resources/mapper/BrandAnalysisMapper.xml
@@ -4,7 +4,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.yupi.springbootinit.mapper.BrandAnalysisMapper">
 
-    <select id="getSpecSalesByBrand" resultType="com.yupi.springbootinit.model.vo.BrandAnalysisResult.SpecSales">
+    <select id="getSpecSalesByBrand" resultType="com.yupi.springbootinit.model.vo.BrandAnalysisResult$SpecSales">
         SELECT p.specification, SUM(co.order_amount) AS totalOrderAmount
         FROM products p
                  JOIN customer_orders co ON p.id = co.product_id
@@ -12,7 +12,7 @@
         GROUP BY p.specification
     </select>
 
-    <select id="getRegionSalesByBrand" resultType="com.yupi.springbootinit.model.vo.BrandAnalysisResult.RegionSales">
+    <select id="getRegionSalesByBrand" resultType="com.yupi.springbootinit.model.vo.BrandAnalysisResult$RegionSales">
         SELECT r.region_name, SUM(co.order_amount) AS totalOrderAmount
         FROM products p
                  JOIN customer_orders co ON p.id = co.product_id
