+++ Gra-frontend/package.json
@@ -97,5 +97,9 @@
   },
   "engines": {
     "node": ">=12.0.0"
-  }
+  },
+  "main": ".eslintrc.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
